fixed invalid PUT & DELETE route for categories

+++ src/app/api/categories/[id]/route.ts
@@ -2,22 +2,33 @@ import { NextRequest, NextResponse } from 'next/server';
 import { corsHeaders, jsonWithCors } from '@/lib/cors';
 import * as categoryService from '@/lib/services/categoryService';
 
-export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
-    const id = parseInt(params.id);
-    if (isNaN(id)) return jsonWithCors({ error: 'Invalid ID' }, 400);
+export async function PUT(req: NextRequest) {
+    const { searchParams } = new URL(req.url);
+    const idParam = searchParams.get('id');
 
-    const body = await req.json();
-    const updated = await categoryService.updateCategory({ ...body, id });
+    if(idParam){
+        const id = parseInt(idParam);
+        if (isNaN(id)) return jsonWithCors({ error: 'Invalid ID' }, 400);
 
-    return jsonWithCors(updated);
+        const body = await req.json();
+        const updated = await categoryService.updateCategory({ ...body, id });
+
+        return jsonWithCors(updated);
+    }
+    return jsonWithCors({ error: 'Invalid ID' }, 400);
 }
 
-export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
-    const id = parseInt(params.id);
-    if (isNaN(id)) return jsonWithCors({ error: 'Invalid ID' }, 400);
+export async function DELETE(req: NextRequest) {
+    const { searchParams } = new URL(req.url);
+    const idParam = searchParams.get('id');
+    if(idParam){
+        const id = parseInt(idParam);
+        if (isNaN(id)) return jsonWithCors({ error: 'Invalid ID' }, 400);
 
-    await categoryService.deleteCategoryById(id);
-    return jsonWithCors({ message: 'Category deleted' });
+        await categoryService.deleteCategoryById(id);
+        return jsonWithCors({ message: 'Category deleted' });
+    }
+    return jsonWithCors({ error: 'category not deleted, no ID provided'}, 400)
 }
 
 // Handle OPTIONS preflight
